requirements.txt
certifi==2026.4.22
charset-normalizer==3.4.7
idna==3.13
linkify-it-py==2.1.0
llvmlite==0.44.0
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
numba==0.61.2
numpy==2.2.6
pandas==3.0.2
pandas-ta==0.4.71b0
platformdirs==4.9.6
plotext==5.3.2
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
rich==15.0.0
schedule==1.2.2
six==1.17.0
textual==8.2.5
textual-plotext==1.0.1
tqdm==4.67.3
typing_extensions==4.15.0
uc-micro-py==2.0.0
urllib3==2.6.3
